Add a RouterBase type
