+++ src/Components/Components/src/Routing/RouterBase.cs
@@ -1,0 +1,127 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Components.Layouts;
+using Microsoft.AspNetCore.Components.RenderTree;
+using Microsoft.AspNetCore.Components.Services;
+
+namespace Microsoft.AspNetCore.Components.Routing
+{
+    /// <summary>
+    /// A base type for custom component routers.
+    /// </summary>
+    public abstract class RouterBase : IComponent, IDisposable
+    {
+        static readonly char[] _queryOrHashStartChar = new[] { '?', '#' };
+
+        RenderHandle _renderHandle;
+        string _baseUri;
+        string _locationAbsolute;
+
+        [Inject] private IUriHelper UriHelper { get; set; }
+
+        /// <summary>
+        /// Gets or sets the type of the component that should be used as a fallback when no match is found for the requested route.
+        /// </summary>
+        [Parameter] private Type FallbackComponent { get; set; }
+
+        private RouteTable Routes { get; set; }
+
+        /// <inheritdoc />
+        public void Configure(RenderHandle renderHandle)
+        {
+            _renderHandle = renderHandle;
+            _baseUri = UriHelper.GetBaseUri();
+            _locationAbsolute = UriHelper.GetAbsoluteUri();
+            UriHelper.OnLocationChanged += OnLocationChanged;
+        }
+
+        /// <inheritdoc />
+        public Task SetParametersAsync(ParameterCollection parameters)
+        {
+            parameters.SetParameterProperties(this);
+            var types = ResolveRoutableComponents();
+            Routes = RouteTable.Create(types);
+            Refresh();
+            return Task.CompletedTask;
+        }
+
+        /// <summary>
+        /// Resolves the types of components that can be routed to.
+        /// </summary>
+        /// <returns>An enumerable of types of components that can be routed to.</returns>
+        protected abstract IEnumerable<Type> ResolveRoutableComponents();
+
+        /// <inheritdoc />
+        public void Dispose()
+        {
+            Dispose(true);
+        }
+
+        /// <summary>
+        /// Frees resources used by the component.
+        /// </summary>
+        /// <param name="disposing">A boolean value indicating whether this is a managed dispose.</param>
+        protected virtual void Dispose(bool disposing)
+        {
+            UriHelper.OnLocationChanged -= OnLocationChanged;
+        }
+
+        private string StringUntilAny(string str, char[] chars)
+        {
+            var firstIndex = str.IndexOfAny(chars);
+            return firstIndex < 0
+                ? str
+                : str.Substring(0, firstIndex);
+        }
+
+        /// <inheritdoc />
+        protected virtual void Render(RenderTreeBuilder builder, Type handler, IDictionary<string, object> parameters)
+        {
+            builder.OpenComponent(0, typeof(LayoutDisplay));
+            builder.AddAttribute(1, LayoutDisplay.NameOfPage, handler);
+            builder.AddAttribute(2, LayoutDisplay.NameOfPageParameters, parameters);
+            builder.CloseComponent();
+        }
+
+        private void Refresh()
+        {
+            var locationPath = UriHelper.ToBaseRelativePath(_baseUri, _locationAbsolute);
+            locationPath = StringUntilAny(locationPath, _queryOrHashStartChar);
+            var context = new RouteContext(locationPath);
+            Routes.Route(context);
+
+            if (context.Handler == null)
+            {
+                if (FallbackComponent != null)
+                {
+                    context.Handler = FallbackComponent;
+                }
+                else
+                {
+                    throw new InvalidOperationException($"'{nameof(Router)}' cannot find any component with a route for '/{locationPath}', and no fallback is defined.");
+                }
+            }
+
+            if (!typeof(IComponent).IsAssignableFrom(context.Handler))
+            {
+                throw new InvalidOperationException($"The type {context.Handler.FullName} " +
+                    $"does not implement {typeof(IComponent).FullName}.");
+            }
+
+            _renderHandle.Render(builder => Render(builder, context.Handler, context.Parameters));
+        }
+
+        private void OnLocationChanged(object sender, string newAbsoluteUri)
+        {
+            _locationAbsolute = newAbsoluteUri;
+            if (_renderHandle.IsInitialized)
+            {
+                Refresh();
+            }
+        }
+    }
+}
